% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-spec.R
\name{all_valid_specs}
\alias{all_valid_specs}
\title{Enumerate all valid model specifications}
\usage{
all_valid_specs()
}
\value{
A tibble with columns: measurement, structural, population, item, link.
}
\description{
Enumerate all valid model specifications
}
