% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-compile.R
\name{clear_stan_cache}
\alias{clear_stan_cache}
\title{Clear the autoskill Stan model cache}
\usage{
clear_stan_cache(dir = NULL)
}
\arguments{
\item{dir}{Cache directory.}
}
\description{
Clear the autoskill Stan model cache
}
