% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-compare.R
\name{compare_models}
\alias{compare_models}
\title{Compare multiple fitted models by LOO-CV}
\usage{
compare_models(...)
}
\arguments{
\item{...}{Named \code{fit_result} objects.}
}
\value{
A tibble with LOO comparison statistics.
}
\description{
Compare multiple fitted models by LOO-CV
}
