% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan-compile.R
\name{compile_model}
\alias{compile_model}
\title{Compile a Stan model from a model configuration}
\usage{
compile_model(config, dir = NULL, force = FALSE)
}
\arguments{
\item{config}{A \code{model_config} object.}

\item{dir}{Cache directory.}

\item{force}{If TRUE, recompile even if cached.}
}
\value{
A CmdStanModel object.
}
\description{
Compile a Stan model from a model configuration
}
