% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{compute_loo}
\alias{compute_loo}
\title{Compute LOO-CV via PSIS-LOO}
\usage{
compute_loo(fit)
}
\arguments{
\item{fit}{A CmdStanMCMC object with log_lik in generated quantities.}
}
\value{
A loo object.
}
\description{
Compute LOO-CV via PSIS-LOO
}
