% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal-problem.R
\name{empty_dag}
\alias{empty_dag}
\title{Construct an empty DAG over named variables}
\usage{
empty_dag(variables)
}
\arguments{
\item{variables}{Character vector of variable names.}
}
\value{
A \code{causal_dag} with no edges.
}
\description{
Construct an empty DAG over named variables
}
