% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{extract_diagnostics}
\alias{extract_diagnostics}
\title{Extract MCMC diagnostics}
\usage{
extract_diagnostics(fit)
}
\arguments{
\item{fit}{A CmdStanMCMC object.}
}
\value{
A tibble with columns: metric, value, status.
}
\description{
Extract MCMC diagnostics
}
