% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-fit.R
\name{extract_param_summary}
\alias{extract_param_summary}
\title{Extract parameter summaries mapped to item/skill names}
\usage{
extract_param_summary(fit, config)
}
\arguments{
\item{fit}{A CmdStanMCMC object.}

\item{config}{A model_config object.}
}
\value{
A tibble with parameter summaries.
}
\description{
Extract parameter summaries mapped to item/skill names
}
