% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-compare.R
\name{flag_problem_items}
\alias{flag_problem_items}
\title{Flag observations with high Pareto k}
\usage{
flag_problem_items(fit_result, k_threshold = 0.7)
}
\arguments{
\item{fit_result}{A \code{fit_result} object.}

\item{k_threshold}{Pareto k threshold (default 0.7).}
}
\value{
A tibble with flagged observations.
}
\description{
Flag observations with high Pareto k
}
