% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-spec.R
\name{is_fa_mode}
\alias{is_fa_mode}
\title{Test whether a model spec uses factor-analytic structural mode}
\usage{
is_fa_mode(spec)
}
\arguments{
\item{spec}{A \code{model_spec} object.}
}
\value{
Logical.
}
\description{
Test whether a model spec uses factor-analytic structural mode
}
