% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-problem.R
\name{is_structure_problem}
\alias{is_structure_problem}
\title{Test whether an object is a \code{structure_problem}}
\usage{
is_structure_problem(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a \code{structure_problem}
}
