% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading-structure.R
\name{loading_structure}
\alias{loading_structure}
\title{Create a loading structure}
\usage{
loading_structure(taxonomy, assignments, items)
}
\arguments{
\item{taxonomy}{A tibble with columns \code{skill_id}, \code{name}, \code{description}.}

\item{assignments}{A tibble with columns \code{item_id}, \code{skill_id}, \code{skill_name}.}

\item{items}{A tibble with columns \code{item_id}, \code{text}.}
}
\value{
An S3 object of class \code{loading_structure}.
}
\description{
Create a loading structure
}
