% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Numerically stable log-sum-exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric scalar \code{log(sum(exp(x)))}.
}
\description{
Numerically stable log-sum-exp
}
