% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-config.R
\name{model_config}
\alias{model_config}
\title{Create a model configuration}
\usage{
model_config(spec, struc, edge_prior = NULL)
}
\arguments{
\item{spec}{A \code{model_spec} object.}

\item{struc}{A \code{loading_structure} object.}

\item{edge_prior}{An \code{edge_prior} object (required for \code{structural = "dag"}).}
}
\value{
An S3 object of class \code{model_config}.
}
\description{
Create a model configuration
}
