% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-spec.R
\name{model_spec}
\alias{model_spec}
\title{Create a model specification}
\usage{
model_spec(
  measurement = "linear",
  structural = "independent",
  population = "single",
  item = "basic",
  link = "logit"
)
}
\arguments{
\item{measurement}{One of "linear" or "interaction".}

\item{structural}{One of "independent", "correlated", "dag", "hierarchical".}

\item{population}{One of "single" or "grouped".}

\item{item}{One of "basic" or "slip_guess".}

\item{link}{One of "logit" or "probit".}
}
\value{
An S3 object of class \code{model_spec}.
}
\description{
Create a model specification
}
