% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{normalize_log_weights}
\alias{normalize_log_weights}
\title{Numerically stable log-sum-exp normalisation}
\usage{
normalize_log_weights(log_weights)
}
\arguments{
\item{log_weights}{Numeric vector.}
}
\value{
Numeric vector the same length as input; \code{sum(exp(result)) == 1}.
}
\description{
Numerically stable log-sum-exp normalisation
}
