% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-data.R
\name{response_data}
\alias{response_data}
\title{Create a response data object}
\usage{
response_data(responses)
}
\arguments{
\item{responses}{A matrix or tibble of student responses.}
}
\value{
An S3 object of class \code{response_data}.
}
\description{
Create a response data object
}
