% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure-optimizer.R
\name{run_iteration}
\alias{run_iteration}
\title{Run one iteration of the optimization loop}
\usage{
run_iteration(
  responses,
  config,
  previous_results = list(),
  skip_sbc = TRUE,
  ...
)
}
\arguments{
\item{responses}{A \code{response_data} object.}

\item{config}{A \code{model_config} object.}

\item{previous_results}{List of previous iteration_results.}

\item{skip_sbc}{If TRUE, skip SBC check.}

\item{...}{Additional arguments passed to \code{fit_model()}.}
}
\value{
An S3 object of class \code{iteration_result}.
}
\description{
Run one iteration of the optimization loop
}
