% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc-check.R
\name{run_sbc}
\alias{run_sbc}
\title{Run simulation-based calibration}
\usage{
run_sbc(config, n_sims = 100, n_students = 200, ...)
}
\arguments{
\item{config}{A \code{model_config} object.}

\item{n_sims}{Number of SBC simulations.}

\item{n_students}{Number of students per simulation.}

\item{...}{Additional arguments passed to sampling.}
}
\value{
An S3 object of class \code{sbc_result}.
}
\description{
Run simulation-based calibration
}
