% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{sbc_generator}
\alias{sbc_generator}
\title{Create an SBC generator function}
\usage{
sbc_generator(config, n_students = 200)
}
\arguments{
\item{config}{A \code{model_config} object.}

\item{n_students}{Number of students per simulation.}
}
\value{
A function that returns a list with \code{variables} and \code{generated}.
}
\description{
Create an SBC generator function
}
