% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Effective sample size of a particle population}
\usage{
effective_sample_size(log_weights)
}
\arguments{
\item{log_weights}{Normalised log-weights.}
}
\value{
Numeric scalar in \verb{[1, length(log_weights)]}.
}
\description{
Effective sample size of a particle population
}
